+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # 2.3.7
 - Add Right-Clicking on Baby Villager With Weakness Potion with a Carved Pumpkin to turn them into a Pumpkin Warden
+- Remove BWGItemTags for Dye Conversions, Replace with Explicit Recipes (Like Vanilla)
+- Fix Tall Flowers only making 1 dye instead of 2
 
 # 2.3.6
 - Fix Vine like blocks not being compostable on Forge